s This is a combination of 2 commits.
Remove inheritance from NormalItem

Inheriting from NormalItem was wrong, and I only did it to share the helper methods, which wasn't a great choice.
Now all Items have their own class without inheriting from NormalItem, since that might bring issues later on when behaviour in NormalItem would change. That would mean we would have to change all code in all types, which is a big no.
I also thought of just removing the helper methods (and with that the shared module) but I made the decision to keep it, since it keeps the shared behavioyr centralised and not scattered around multiple models.
Another thought that came back was changing the quality in one go instead of in multiple steps. I still kept it for verbosity and because the changes are so light/minimal in this exercise. I would not keep it like this if the change was something heavier than changing an integer or if every change had follow up actions.

push concern

undo concern change